Added the basic tabs
BUT there are lots of bugs to work out
