+++ AndroidProject/DeLighted/app/src/main/java/com/example/caballero/delighted/TabThreeFragment.java
@@ -1,0 +1,19 @@
+package com.example.caballero.delighted;
+
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+public class TabThreeFragment extends android.support.v4.app.Fragment{
+    private static final String TAG = "TabThreeFragment";
+
+    @Nullable
+    @Override
+    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        View view = inflater.inflate(R.layout.tab3_frag,container,false);
+
+        return view;
+    }
+}
