+++ AndroidProject/DeLighted/app/src/main/res/layout/tab2_frag.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<android.support.constraint.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <TextView
+        android:id="@+id/tabtwotext"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="Hello there tab two user"
+        tools:layout_editor_absoluteX="66dp"
+        tools:layout_editor_absoluteY="32dp" />
+</android.support.constraint.ConstraintLayout>
